consolidate operations.py into base.py
